requirements.txt
unicodedata
requests
snowballstemmer
pyenchant
nltk
scikit-learn
seaborn
bs4
pandas
numpy